refactor: moved lerp back to Rocket
